+++ openmls/src/group/mls_group/builder.rs
@@ -2,16 +2,16 @@ use openmls_traits::{signatures::Signer, types::Ciphersuite};
 use tls_codec::Serialize;
 
 use crate::{
-    binary_tree::array_representation::TreeSize,
+    binary_tree::{array_representation::TreeSize, LeafNodeIndex},
     credentials::CredentialWithKey,
     error::LibraryError,
     extensions::{errors::InvalidExtensionError, Extensions},
     group::{
-        public_group::errors::PublicGroupBuildError, GroupId, MlsGroupCreateConfig,
-        MlsGroupCreateConfigBuilder, NewGroupError, PublicGroup, WireFormatPolicy,
+        past_secrets::MessageSecretsStore, public_group::errors::PublicGroupBuildError, GroupId,
+        MlsGroup, MlsGroupCreateConfig, MlsGroupCreateConfigBuilder, MlsGroupState, NewGroupError,
+        PublicGroup, WireFormatPolicy,
     },
     key_packages::Lifetime,
-    prelude::LeafNodeIndex,
     schedule::{
         psk::{load_psks, store::ResumptionPskStore, PskSecret},
         InitSecret, JoinerSecret, KeySchedule, PreSharedKeyId,
@@ -20,8 +20,6 @@ use crate::{
     tree::sender_ratchet::SenderRatchetConfiguration,
     treesync::{errors::LeafNodeValidationError, node::leaf_node::Capabilities},
 };
-
-use super::{past_secrets::MessageSecretsStore, MlsGroup, MlsGroupState};
 
 /// Builder struct for an [`MlsGroup`].
 #[derive(Default, Debug)]