+++ openmls/src/group/tests_and_kats/tests/capabilities_check.rs
@@ -154,6 +154,7 @@ impl<'a, 'b: 'a, Provider: OpenMlsProvider> PreGroupPartyState<'b, Provider> {
             },
             Extensions::default(),
             &self.core_state.provider,
+            None,
             &self.signer,
         );
     }